+++ app/(pages)/(hackers)/_components/ProjectInfo/FAQAccordian/MusicPlayer.tsx
@@ -3,7 +3,7 @@
 import { useState, useRef, useEffect } from 'react';
 import Image from 'next/image';
 import styles from './MusicPlayer.module.scss';
-import vinyl from 'public/judges/hub/vinyl.svg';
+import vinyl from 'public/hackers/vinyl.svg';
 import pauseIcon from 'public/hackers/pause-icon.svg';
 import playIcon from 'public/hackers/play-icon.svg';
